Remove old interface
No changes to UJSJobState other than rename

Also some test fixes
