+++ src/us/kbase/userandjobstate/UserAndJobStateServer.java
@@ -58,7 +58,7 @@ import us.kbase.common.mongo.exceptions.MongoAuthException;
 import us.kbase.userandjobstate.jobstate.Job;
 import us.kbase.userandjobstate.jobstate.JobResult;
 import us.kbase.userandjobstate.jobstate.JobResults;
-import us.kbase.userandjobstate.jobstate.UJSJobState;
+import us.kbase.userandjobstate.jobstate.JobState;
 import us.kbase.userandjobstate.userstate.UserState;
 import us.kbase.userandjobstate.userstate.UserState.KeyState;
 //END_HEADER
@@ -149,7 +149,7 @@ public class UserAndJobStateServer extends JsonServerServlet {
 	private final static int TOKEN_REFRESH_INTERVAL_SEC = 24 * 60 * 60;
 	
 	private final UserState us;
-	private final UJSJobState js;
+	private final JobState js;
 	private final ConfigurableAuthService auth;
 	
 	private final static DateTimeFormatter DATE_PARSER =
@@ -229,10 +229,10 @@ public class UserAndJobStateServer extends JsonServerServlet {
 		return null;
 	}
 	
-	private UJSJobState getJobState(final DB db, final SchemaManager sm,
+	private JobState getJobState(final DB db, final SchemaManager sm,
 			final String host) {
 		try {
-			return new UJSJobState(db.getCollection(JOB_COLLECTION), sm);
+			return new JobState(db.getCollection(JOB_COLLECTION), sm);
 		} catch (MongoTimeoutException e) {
 			fail("Couldn't connect to mongo host " + host + ": " +
 					e.getLocalizedMessage());
@@ -760,15 +760,15 @@ public class UserAndJobStateServer extends JsonServerServlet {
 		if (progress.getPtype() == null) {
 			throw new IllegalArgumentException("Progress type cannot be null");
 		}
-		if (progress.getPtype().equals(UJSJobState.PROG_NONE)) {
+		if (progress.getPtype().equals(JobState.PROG_NONE)) {
 			js.startJob(authPart.getUserName(), job,
 					getServiceUserName(token), status, desc,
 					parseDate(estComplete));
-		} else if (progress.getPtype().equals(UJSJobState.PROG_PERC)) {
+		} else if (progress.getPtype().equals(JobState.PROG_PERC)) {
 			js.startJobWithPercentProg(
 					authPart.getUserName(), job, getServiceUserName(token), status,
 					desc, parseDate(estComplete));
-		} else if (progress.getPtype().equals(UJSJobState.PROG_TASK)) {
+		} else if (progress.getPtype().equals(JobState.PROG_TASK)) {
 			if (progress.getMax() == null) {
 				throw new IllegalArgumentException(
 						"Max progress cannot be null for task based progress");
@@ -813,14 +813,14 @@ public class UserAndJobStateServer extends JsonServerServlet {
 		if (progress.getPtype() == null) {
 			throw new IllegalArgumentException("Progress type cannot be null");
 		}
-		if (progress.getPtype().equals(UJSJobState.PROG_NONE)) {
+		if (progress.getPtype().equals(JobState.PROG_NONE)) {
 			returnVal = js.createAndStartJob(authPart.getUserName(),
 					getServiceUserName(token), status, desc, parseDate(estComplete));
-		} else if (progress.getPtype().equals(UJSJobState.PROG_PERC)) {
+		} else if (progress.getPtype().equals(JobState.PROG_PERC)) {
 			returnVal = js.createAndStartJobWithPercentProg(
 					authPart.getUserName(), getServiceUserName(token), status, desc,
 					parseDate(estComplete));
-		} else if (progress.getPtype().equals(UJSJobState.PROG_TASK)) {
+		} else if (progress.getPtype().equals(JobState.PROG_TASK)) {
 			if (progress.getMax() == null) {
 				throw new IllegalArgumentException(
 						"Max progress cannot be null for task based progress");
